# Optimize intermediateNodeDB.constructDBKey